Network.AWS.Shield

Description

AWS Shield Advanced

This is the AWS Shield Advanced API Reference . This guide is for developers who need detailed information about the AWS Shield Advanced API actions, data types, and errors. For detailed information about AWS WAF and AWS Shield Advanced features and an overview of how to use the AWS WAF and AWS Shield Advanced APIs, see the AWS WAF and AWS Shield Developer Guide .

Service Configuration

shield :: Service

API version 2016-06-02 of the Amazon Shield SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Shield.

InvalidResourceException

_InvalidResourceException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the resource is invalid. You might not have access to the resource, or the resource might not exist.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the parameters passed to the API are invalid.

LimitsExceededException

_LimitsExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the operation would exceed a limit.

InternalErrorException

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that a problem occurred with the service infrastructure. You can retry the request.

ResourceAlreadyExistsException

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

Exception indicating the specified resource already exists.

OptimisticLockException

_OptimisticLockException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the protection state has been modified by another client. You can retry the request.

InvalidOperationException

_InvalidOperationException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the operation would not cause any change to occur.

LockedSubscriptionException

_LockedSubscriptionException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the subscription has been modified by another client. You can retry the request.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Exception indicating the specified resource does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreateSubscription

module Network.AWS.Shield.CreateSubscription

ListProtections

module Network.AWS.Shield.ListProtections

DeleteSubscription

module Network.AWS.Shield.DeleteSubscription

DescribeAttack

module Network.AWS.Shield.DescribeAttack

DescribeProtection

module Network.AWS.Shield.DescribeProtection

ListAttacks

module Network.AWS.Shield.ListAttacks

CreateProtection

module Network.AWS.Shield.CreateProtection

DeleteProtection

module Network.AWS.Shield.DeleteProtection

DescribeSubscription

module Network.AWS.Shield.DescribeSubscription

Types

SubResourceType

data SubResourceType

Constructors

IP

URL

AttackDetail

data AttackDetail

The details of a DDoS attack.

See: attackDetail smart constructor.

attackDetail :: AttackDetail

Creates a value of AttackDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

adAttackId - The unique identifier (ID) of the attack.

adStartTime - The time the attack started, in the format 2016-12-16T13:50Z.

adSubResources - If applicable, additional detail about the resource being attacked, for example, IP address or URL.

adMitigations - List of mitigation actions taken for the attack.

adAttackCounters - List of counters that describe the attack for the specified time period.

adResourceARN - The ARN (Amazon Resource Name) of the resource that was attacked.

adEndTime - The time the attack ended, in the format 2016-12-16T13:50Z.

adAttackId :: Lens' AttackDetail (Maybe Text)

The unique identifier (ID) of the attack.

adStartTime :: Lens' AttackDetail (Maybe UTCTime)

The time the attack started, in the format 2016-12-16T13:50Z.

adSubResources :: Lens' AttackDetail [SubResourceSummary]

If applicable, additional detail about the resource being attacked, for example, IP address or URL.

adMitigations :: Lens' AttackDetail [Mitigation]

List of mitigation actions taken for the attack.

adAttackCounters :: Lens' AttackDetail [SummarizedCounter]

List of counters that describe the attack for the specified time period.

adResourceARN :: Lens' AttackDetail (Maybe Text)

The ARN (Amazon Resource Name) of the resource that was attacked.

adEndTime :: Lens' AttackDetail (Maybe UTCTime)

The time the attack ended, in the format 2016-12-16T13:50Z.

AttackSummary

data AttackSummary

Summarizes all DDoS attacks for a specified time period.

See: attackSummary smart constructor.

attackSummary :: AttackSummary

Creates a value of AttackSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asAttackVectors - The list of attacks for a specified time period.

asAttackId - The unique identifier (ID) of the attack.

asStartTime - The start time of the attack, in the format 2016-12-16T13:50Z.

asResourceARN - The ARN (Amazon Resource Name) of the resource that was attacked.

asEndTime - The end time of the attack, in the format 2016-12-16T13:50Z.

asAttackVectors :: Lens' AttackSummary [AttackVectorDescription]

The list of attacks for a specified time period.

asAttackId :: Lens' AttackSummary (Maybe Text)

The unique identifier (ID) of the attack.

asStartTime :: Lens' AttackSummary (Maybe UTCTime)

The start time of the attack, in the format 2016-12-16T13:50Z.

asResourceARN :: Lens' AttackSummary (Maybe Text)

The ARN (Amazon Resource Name) of the resource that was attacked.

asEndTime :: Lens' AttackSummary (Maybe UTCTime)

The end time of the attack, in the format 2016-12-16T13:50Z.

AttackVectorDescription

data AttackVectorDescription

Describes the attack.

See: attackVectorDescription smart constructor.

attackVectorDescription

Arguments

:: Text
avdVectorType

-> AttackVectorDescription

Creates a value of AttackVectorDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avdVectorType - The attack type, for example, SNMP reflection or SYN flood.

avdVectorType :: Lens' AttackVectorDescription Text

The attack type, for example, SNMP reflection or SYN flood.

Mitigation

data Mitigation

The mitigation applied to a DDoS attack.

See: mitigation smart constructor.

mitigation :: Mitigation

Creates a value of Mitigation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mMitigationName - The name of the mitigation taken for this attack.

mMitigationName :: Lens' Mitigation (Maybe Text)

The name of the mitigation taken for this attack.

Protection

data Protection

An object that represents a resource that is under DDoS protection.

See: protection smart constructor.

protection :: Protection

Creates a value of Protection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pResourceARN - The ARN (Amazon Resource Name) of the AWS resource that is protected.

pName - The friendly name of the protection. For example, My CloudFront distributions .

pId - The unique identifier (ID) of the protection.

pResourceARN :: Lens' Protection (Maybe Text)

The ARN (Amazon Resource Name) of the AWS resource that is protected.

pName :: Lens' Protection (Maybe Text)

The friendly name of the protection. For example, My CloudFront distributions .

pId :: Lens' Protection (Maybe Text)

The unique identifier (ID) of the protection.

SubResourceSummary

data SubResourceSummary

The attack information for the specified SubResource.

See: subResourceSummary smart constructor.

subResourceSummary :: SubResourceSummary

Creates a value of SubResourceSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsCounters - The counters that describe the details of the attack.

srsAttackVectors - The list of attack types and associated counters.

srsId - The unique identifier (ID) of the SubResource .

srsType - The SubResource type.

srsCounters :: Lens' SubResourceSummary [SummarizedCounter]

The counters that describe the details of the attack.

srsAttackVectors :: Lens' SubResourceSummary [SummarizedAttackVector]

The list of attack types and associated counters.

srsId :: Lens' SubResourceSummary (Maybe Text)

The unique identifier (ID) of the SubResource .

srsType :: Lens' SubResourceSummary (Maybe SubResourceType)

The SubResource type.

Subscription

data Subscription

Information about the AWS Shield Advanced subscription for an account.

See: subscription smart constructor.

subscription :: Subscription

Creates a value of Subscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sTimeCommitmentInSeconds - The length, in seconds, of the AWS Shield Advanced subscription for the account.

sStartTime - The start time of the subscription, in the format "2016-12-16T13:50Z".

sTimeCommitmentInSeconds :: Lens' Subscription (Maybe Natural)

The length, in seconds, of the AWS Shield Advanced subscription for the account.

sStartTime :: Lens' Subscription (Maybe UTCTime)

The start time of the subscription, in the format "2016-12-16T13:50Z".

SummarizedAttackVector

data SummarizedAttackVector

A summary of information about the attack.

See: summarizedAttackVector smart constructor.

summarizedAttackVector

Arguments

:: Text
savVectorType

-> SummarizedAttackVector

Creates a value of SummarizedAttackVector with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

savVectorCounters - The list of counters that describe the details of the attack.

savVectorType - The attack type, for example, SNMP reflection or SYN flood.

savVectorCounters :: Lens' SummarizedAttackVector [SummarizedCounter]

The list of counters that describe the details of the attack.

savVectorType :: Lens' SummarizedAttackVector Text

The attack type, for example, SNMP reflection or SYN flood.

SummarizedCounter

data SummarizedCounter

The counter that describes a DDoS attack.

See: summarizedCounter smart constructor.

summarizedCounter :: SummarizedCounter

Creates a value of SummarizedCounter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

scMax - The maximum value of the counter for a specified time period.

scAverage - The average value of the counter for a specified time period.

scN - The number of counters for a specified time period.

scName - The counter name.

scSum - The total of counter values for a specified time period.

scUnit - The unit of the counters.

scMax :: Lens' SummarizedCounter (Maybe Double)

The maximum value of the counter for a specified time period.

scAverage :: Lens' SummarizedCounter (Maybe Double)

The average value of the counter for a specified time period.

scN :: Lens' SummarizedCounter (Maybe Int)

The number of counters for a specified time period.

scName :: Lens' SummarizedCounter (Maybe Text)

The counter name.

scSum :: Lens' SummarizedCounter (Maybe Double)

The total of counter values for a specified time period.

scUnit :: Lens' SummarizedCounter (Maybe Text)

The unit of the counters.

TimeRange

data TimeRange

The time range.

See: timeRange smart constructor.

timeRange :: TimeRange

Creates a value of TimeRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trFromInclusive - The start time, in the format 2016-12-16T13:50Z.

trToExclusive - The end time, in the format 2016-12-16T15:50Z.

trFromInclusive :: Lens' TimeRange (Maybe UTCTime)

The start time, in the format 2016-12-16T13:50Z.

trToExclusive :: Lens' TimeRange (Maybe UTCTime)

The end time, in the format 2016-12-16T15:50Z.