Network.AWS.Shield.ListAttacks

Description

Returns all ongoing DDoS attacks or all DDoS attacks during a specified time period.

Creating a Request

listAttacks :: ListAttacks

Creates a value of ListAttacks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laStartTime - The time period for the attacks.

laResourceARNs - The ARN (Amazon Resource Name) of the resource that was attacked. If this is left blank, all applicable resources for this account will be included.

laNextToken - The ListAttacksRequest.NextMarker value from a previous call to ListAttacksRequest . Pass null if this is the first call.

laEndTime - The end of the time period for the attacks.

laMaxResults - The maximum number of AttackSummary objects to be returned. If this is left blank, the first 20 results will be returned.

data ListAttacks

See: listAttacks smart constructor.

Request Lenses

laStartTime :: Lens' ListAttacks (Maybe TimeRange)

The time period for the attacks.

laResourceARNs :: Lens' ListAttacks [Text]

The ARN (Amazon Resource Name) of the resource that was attacked. If this is left blank, all applicable resources for this account will be included.

laNextToken :: Lens' ListAttacks (Maybe Text)

The ListAttacksRequest.NextMarker value from a previous call to ListAttacksRequest . Pass null if this is the first call.

laEndTime :: Lens' ListAttacks (Maybe TimeRange)

The end of the time period for the attacks.

laMaxResults :: Lens' ListAttacks (Maybe Natural)

The maximum number of AttackSummary objects to be returned. If this is left blank, the first 20 results will be returned.

Destructuring the Response

listAttacksResponse

Arguments

:: Int
larsResponseStatus

-> ListAttacksResponse

Creates a value of ListAttacksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larsAttackSummaries - The attack information for the specified time range.

larsNextToken - The token returned by a previous call to indicate that there is more data available. If not null, more results are available. Pass this value for the NextMarker parameter in a subsequent call to ListAttacks to retrieve the next set of items.

larsResponseStatus - -- | The response status code.

data ListAttacksResponse

See: listAttacksResponse smart constructor.

Response Lenses

larsAttackSummaries :: Lens' ListAttacksResponse [AttackSummary]

The attack information for the specified time range.

larsNextToken :: Lens' ListAttacksResponse (Maybe Text)

The token returned by a previous call to indicate that there is more data available. If not null, more results are available. Pass this value for the NextMarker parameter in a subsequent call to ListAttacks to retrieve the next set of items.

larsResponseStatus :: Lens' ListAttacksResponse Int

- | The response status code.