Network.AWS.Shield.ListProtections

Description

Lists all Protection objects for the account.

Creating a Request

listProtections :: ListProtections

Creates a value of ListProtections with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpNextToken - The ListProtectionsRequest.NextToken value from a previous call to ListProtections . Pass null if this is the first call.

lpMaxResults - The maximum number of Protection objects to be returned. If this is left blank the first 20 results will be returned.

data ListProtections

See: listProtections smart constructor.

Request Lenses

lpNextToken :: Lens' ListProtections (Maybe Text)

The ListProtectionsRequest.NextToken value from a previous call to ListProtections . Pass null if this is the first call.

lpMaxResults :: Lens' ListProtections (Maybe Natural)

The maximum number of Protection objects to be returned. If this is left blank the first 20 results will be returned.

Destructuring the Response

listProtectionsResponse

Arguments

:: Int
lprsResponseStatus

-> ListProtectionsResponse

Creates a value of ListProtectionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsProtections - The array of enabled Protection objects.

lprsNextToken - If you specify a value for MaxResults and you have more Protections than the value of MaxResults, AWS Shield Advanced returns a NextToken value in the response that allows you to list another group of Protections. For the second and subsequent ListProtections requests, specify the value of NextToken from the previous response to get information about another batch of Protections.

lprsResponseStatus - -- | The response status code.

data ListProtectionsResponse

See: listProtectionsResponse smart constructor.

Response Lenses

lprsProtections :: Lens' ListProtectionsResponse [Protection]

The array of enabled Protection objects.

lprsNextToken :: Lens' ListProtectionsResponse (Maybe Text)

If you specify a value for MaxResults and you have more Protections than the value of MaxResults, AWS Shield Advanced returns a NextToken value in the response that allows you to list another group of Protections. For the second and subsequent ListProtections requests, specify the value of NextToken from the previous response to get information about another batch of Protections.

lprsResponseStatus :: Lens' ListProtectionsResponse Int

- | The response status code.